Updated install scripts.

+++ deps.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 # Easy installer script for Debian/RedHat systems.
+# TODO: PyOpenGL not installed for Python3.
 
 SUDO=$(which sudo)
 SUMOUNT="$1 $2"
@@ -7,10 +8,24 @@ SUMOUNT="$1 $2"
 function fmk
 {
 	# Get and build the firmware mod kit
-	$SUDO rm -rf /opt/firmware-mod-kit/
-	$SUDO mkdir -p /opt/firmware-mod-kit
-	$SUDO chmod a+rwx /opt/firmware-mod-kit
-	git clone https://code.google.com/p/firmware-mod-kit /opt/firmware-mod-kit/
+	if [ -e /opt/firmware-mod-kit ]
+	then
+		if [ ! -e /opt/firmware-mod-kit/.git ]
+		then
+			$SUDO rm -rf /opt/firmware-mod-kit/
+		fi
+	fi
+
+	if [ ! -e /opt/firmware-mod-kit ]
+	then
+		$SUDO mkdir -p /opt/firmware-mod-kit
+		$SUDO chmod a+rwx /opt/firmware-mod-kit
+	fi
+
+	if [ ! -e /opt/firmware-mod-kit/.git ]
+	then
+		git clone https://code.google.com/p/firmware-mod-kit /opt/firmware-mod-kit/
+	fi
 
 	cd /opt/firmware-mod-kit/src
 	./configure && $SUDO make
@@ -35,28 +50,6 @@ function fmk
 	cd -
 }
 
-#function libmagic
-#{
-#	SITE="ftp://ftp.astron.com/pub/file/"
-#	VERSION="5.11"
-#	OUTFILE="file-$VERSION.tar.gz"
-#	URL="$SITE$OUTFILE"
-#
-#	echo "Downloading '$URL'..."
-#	wget "$URL"
-#
-#	if [ -e "$OUTFILE" ]
-#	then
-#		echo "Installing libmagic / python-magic..."
-#		tar -zxvf "$OUTFILE"
-#		cd "file-$VERSION" && ./configure && make && $SUDO make install && cd python && $SUDO python ./setup.py install && cd ../..
-#		$SUDO rm -rf "file-$VERSION" "$OUTFILE"
-#	else
-#		echo "ERROR: Failed to download '$URL'!"
-#		echo "libmagic not installed."
-#	fi
-#}
-
 function pyqtgraph
 {
 	SITE="http://www.pyqtgraph.org/downloads/"
@@ -64,6 +57,8 @@ function pyqtgraph
 	OUTFILE="pyqtgraph-$VERSION.tar.gz"
 	URL="$SITE$OUTFILE"
 
+	pyopengl
+	
 	echo "Downloading '$URL'..."
 	wget "$URL"
 
@@ -72,6 +67,10 @@ function pyqtgraph
 		echo "Installing pyqtgraph..."
 		tar -zxvf "$OUTFILE"
 		cd "pyqtgraph-$VERSION" && $SUDO python ./setup.py install && cd ..
+		if [ "$(which python3)" != "" ]
+		then
+			cd "pyqtgraph-$VERSION" && $SUDO python3 ./setup.py install && cd ..
+		fi
 		$SUDO rm -rf "pyqtgraph-$VERSION" "$OUTFILE"
 	else
 		echo "ERROR: Failed to download '$URL'!"
@@ -90,35 +89,27 @@ function debian
 	fi
 
 	# Install binwalk/fmk pre-requisites and extraction tools
-	$SUDO apt-get -y install git build-essential mtd-utils zlib1g-dev liblzma-dev ncompress gzip bzip2 tar arj p7zip p7zip-full openjdk-6-jdk
-	$SUDO apt-get -y install python-opengl python-qt4 python-qt4-gl python-numpy python-scipy
+	$SUDO apt-get -y install git build-essential mtd-utils zlib1g-dev liblzma-dev ncompress gzip bzip2 tar arj p7zip p7zip-full openjdk-6-jdk libfuzzy2
+	$SUDO apt-get -y install libqt4-opengl python-opengl python-qt4 python-qt4-gl python-numpy python-scipy
 	if [ "$(which python3)" != "" ]
 	then
 		$SUDO apt-get -y install python3-pyqt4 python3-numpy python3-scipy
-		$SUDO apt-get -y install python3-pip
-		PIP="$(which pip-3* | head -1)"
-		if [ "$PIP" != "" ]
-		then
-			$SUDO $PIP install PyOpenGl
-		fi
 	fi
+
+	pyqtgraph
 }
 
 function redhat
 {
 	$SUDO yum groupinstall -y "Development Tools"
-	$SUDO yum install -y git mtd-utils unrar zlib1g-dev liblzma-dev xz-devel compress gzip bzip2 tar arj p7zip p7zip-full openjdk-6-jdk
-	$SUDO yum install -y python-opengl python-qt4 python-qt4-gl python-numpy python-scipy
+	$SUDO yum install -y git mtd-utils unrar zlib1g-dev liblzma-dev xz-devel compress gzip bzip2 tar arj p7zip p7zip-full openjdk-6-jdk libfuzzy2
+	$SUDO yum install -y libqt4-opengl python-opengl python-qt4 python-qt4-gl python-numpy python-scipy
 	if [ "$(which python3)" != "" ]
 	then
 		$SUDO yum -y install python3-pyqt4 python3-numpy python3-scipy
-		$SUDO yum -y install python3-pip
-		PIP="$(which pip-3* | head -1)"
-		if [ "$PIP" != "" ]
-		then
-			$SUDO $PIP install PyOpenGl
-		fi
 	fi
+	
+	pyqtgraph
 }
 
 if [ "$1" == "" ] || [ "$1" == "--sumount" ]
@@ -176,18 +167,6 @@ case $DISTRO in
 		exit 1
 esac
 
-#if [ "$(python -c 'import magic; print (magic.MAGIC_NO_CHECK_TEXT)' 2>/dev/null)" == "" ]
-#then
-#	echo "python-magic not installed, or wrong version; building from source..."
-#	libmagic
-#fi
-
 # Get and build the firmware mod kit
 fmk
-
-if [ "$(python -c 'import pyqtgraph; print (pyqtgraph.__file__)' 2>/dev/null)" == "" ]
-then
-	echo "pyqtgraph not installed; building from source..."
-	pyqtgraph
-fi
